Move payment creation to the service
